[BUGFIX] Add default value for language in flex inline child table
The language value is not displayed in BE for editing, and the missing
default value causes it to be empty upon inserting to DB. The field
however is defined as integer.
This causes the insert operation to fail.
By providing a default language value, the requirements for insert
to DB are fulfilled.