[TASK] styleguide: Add additional autoSizeMax examples for type group and inline
The option autoSizeMax is a common property between select, group and
inline fields. This patch adds examples for group and inline fields.

Inline fields have to fulfill special conditions for autoSizeMax
to be useful:
1. foreign_selector has to be defined.
2. The field, foreign_selector is pointing to, has to be of type select.

Only then a select field is rendered, which autoSizeMax can be applied
to.

Also increase the minimum value to at least 10 or else no effect will be
visible because of a css min-height of 156px in TYPO3 v11. This equals
~7 rows depending on the browser.