[TASK] Avoid deprecated phpstan configuration 'excludes_analyse'
PHPStan configuration option 'excludes_analyse' is deprecated and
should be replaced with the new option 'excludePaths' which would
prepare for PHPStan version raise. At least this suppress the
deprecation notice on phpstan execution.

Added a todo comment for option 'checkMissingIterableValueType'
why this is needed and when to recheck and remove along the way.

Releases: main, 11