[BUGFIX] Use valid be_session fixture to ensure acceptance tests works
There has been quite some changes regarding the session system and on the
database table structure of be_sessions, additionally the provided fixture
file in testing_framework has changed to be compatible with changes in v11.
Because of the fact testing_framework version 6.x is used for v10 and v11
testing, this fixture no longer works for acceptance testing against v10 core.

This change copies the provided be_session.xml fixture from 10.4-dev branch
to this extension and use this local fixture instead of the one from the
testing_framework.

This makes commit and pre-merge testing green on styleguide 10 branch.