[TASK] Remove tableWizard example (#289)
The tableWizard as a fieldControl for the textTable element in gone in
TYPO3 v11:
https://review.typo3.org/c/Packages/TYPO3.CMS/+/70823

Instead, it is always displayed inline. The real textarea field is now
hidden from the user.

Because of that, setting "rows" doesn't make any sense anymore, besides
when using together with "readOnly". Thus, it is removed from the
example.

Along with that "cols", "numNewRows" and "xmlOutput" are no longer
configurable.

The removed field text_21 is not used in TYPO3 acceptance tests, so it
needs no adjustments from that side.

Releases: main, 11.5