[TASK] Use phpstan github error format in github actions
Use specific github error formatting for phpstan
error reporing in github actions. This adds the
phpstan errors directly into the file view on pull
requests for better readability. For direct pushes
on commit level it directly extracts and shows the
errors in the pane instead of only in the command
line view.

Releases: main, 11, 10