[BUGFIX] styleguide: Ensure testing works with CMSComposerInstallers 5
TYPY3v12 recently added hard requirement for `typo3/cms-composer-installers:5.x`,
which install core system extensions and extensions in a different way. This
breaks how typo3/testing-framework expects the existing structure to work for
unit, functional and acceptance tests.

Until the final overwork for the testing-structure is implemented, the
intermediate solution in form of a composer plugin is used, which basicly
provides through symlinks the expected structure.

This *must* and will be vanish in the future or replaced by the official way.

Used command(s):

```shell
  composer require --dev \
    sbuerk/typo3-cmscomposerinstallers-testingframework-bridge:^0.0.1
```