Allow creation of plugins and add indexed_search
For easy debugging of the indexed search extension I would like to add a page with the plugin. Similar things are done for felogin and form. But the indexed_search isn't a dedicated CType (yet!?). Therefore I've implemented the creation of plugins in general and added the indexed search plugin a the first example.