[TASK] Compatibility with associated array keys for TCA items
This is a preparation patch for the upcoming switch to associated items
array keys. The default value generators for select and radio are
adjusted to work for both variants, so tests for the mentioned patch
can go green.

Related patch: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77626

Releases: main