[TASK] styleguide: Replace obsolete Font Awesome icons
This change replaces the obsolete "Font Awesome Icons"
markup on the icon search page and in the section to
create example page trees with TCA demo records and
content elements.

JavaScript selectors are adjusted to match the changed
markup.

The spinner is changed from dark as a side-change for
a cleaner user experiance and aligns to a recently core
patch [1].

[1] https://review.typo3.org/c/Packages/TYPO3.CMS/+/80103

Resolves: https://github.com/TYPO3-CMS/styleguide/issues/416
Releases: main, 12