[FEATURE] Enable item groups for foreign_table
It's now possible to define a specific field
for a foreign table, holding the id of an
associated item group. This therefore enables
the item groups feature also for select fields,
fetching its items from a foreign table.

The new option is called `foreign_table_item_group`.
In case an item group of a foreign item is not
defined in the `itemsGroup` config of the local
field, the group id is used as label. This is the
same as if a static item contains a not configured
group.

Resolves: #95808
Related: #91008
Releases: main
Change-Id: I111d5c1bd4507347a3a7e15e9779d010580838dd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82171
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>