[TASK] Doctrine DBAL 3.2 Connection::lastInsertId() mitigation
The usage of `Connection::lastInsertId()` with a sequence name
is deprecated as unsafe in scenarios with multiple concurrent
connections since Doctrine DBAL 3.2. If a newly inserted row
needs to be referenced, it is recommended to generate its
identifier explicitly prior to insertion.

This change adjusts the usages to avoid calling the method with
the `$name` and/or the TYPO3 additional `$fieldName` argument.

The method signature is not adjusted yet, due to the bounding
to the extended connection method signature.

[1] https://github.com/doctrine/dbal/blob/3.8.x/UPGRADE.md#deprecated-connectionlastinsertidname

Resolves: #102863
Releases: main
Change-Id: I499104d0048584936b7ad2771b5cc977bf0372a7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82511
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>