[BUGFIX] Respect TCA type `group` as `foreign_selector`
It's now possible to use a TCA type `group`
field as `foreign_selector` together with
the `useCombination` functionality. The
corresponding "child child table" information
is in this case fetched from the `allowed`
option.

Side note: Using TCA type `group` for a
`foreign_selector` field is not really
documented, but it's also not forbidden.
Actually, it's a valid use case and also
working quite well as the new styleguide
examples demonstrate.

However, the documentation should be adjusted
to mention that only one table can be defined
in `allowed` for such fields.

Resolves: #102904
Releases: main, 12.4
Change-Id: Icdde6a8acb1fb9166da8e6259758f6bf64d6c41f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82721
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>