[BUGFIX] Respect option 'with site set' for ext:styleguide generator
With #103523 site sets for ext:styleguide were introduced and
the page tree generator received an option to create a frontend
demo section using the shipped set.

Now the click on this button will actually respect the user's wish
and create the site using the set and not create a sys_template
record.

Resolves: #103732
Related: #103523
Releases: main
Change-Id: I1560150c24f643a9ac1b7aefe3d7b0fbe109d62e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84034
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>