[BUGFIX] Prevent multiple FlexContainers of same name
It is possible in Flexform containers to set a XML
configuration that uses the same input field names,
but within different sheets.

The `FlexFormSectionContainer::render()` method used a
HTML DOM id attribute that did not consider
the Flexform sheet name, which has lead to the same id
attribute being re-used. The JavaScript events (add
and delete items from a container) would then
fire only on the first instance, but as many times
as there were wrongly regarded "duplicate" containers
using the same id.

This fix introduces another identifier section to the
id attribute, containing the Flexform sheet name.
This XML attribute is compatible
with allowed CSS class names, so no further escaping is
done with it, similar to the use of `$flexFormFieldName`.

Resolves: #97042
Related: #99947
Releases: main, 12.4
Change-Id: Ic1c81c658b090cb7b9c1a4230c0b4e75c12560dc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85083
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>