[TASK] Add "overrideChildTCA" cases to EXT:styleguide
EXT:styleguide did not contain a showcase for a overrideChildTCA
configuration using both columnOverrides or columns to turn
for example a "plaintext" input element to "RTE".

This case is helpful to also reproduce problems with
FormEngine/DataHandler, where possible RTE transformations on
save might not be calculated when used as an override.

The new distinct "Form engine - inline with child overrideChildTca"
type has been added, and "Form engine - file" has a new
field "file_6" to showcase an RTE within a type=file inline
reference.

Resolves: #81893
Releases: main
Change-Id: Id36fff2dbb92d84882fed309e029ac465128e6ea
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85148
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>