[TASK] Instantiate FieldValueGenerators using DI
This patch seeks to inject a list of all FieldGenerators instances
in EXT:styleguide instead of creating them using GU::makeInstance()
each time.

To achieve this, the class list is moved from a class property
to a constructor argument in the `Services.yaml`, reverting the
public exposure of the many Generators in the DI container.

This leads to circular dependencies where data generators need
instances of RecordData or FieldGeneratorResolver again. This
is resolved with two new interfaces to set these instances at
runtime.

TableHandler implementations follow the same pattern.

Resolves: #104574
Releases: main
Change-Id: I552e962af4e145d284f73fb8d0d116e8d49dcd5e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85444
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>