[!!!][TASK] Use DB-level GUID Type for TCA type=uuid
This change modifies `DefaultTcaSchema` to set the
Doctrine DBAL Type `GUID` for `TCAtype=uuid` fields,
using native `UUID` in case of PostgresSql and the
fallback `CHAR(36)` for other dbms.

It's technically debatable if this is really breaking
and the decision to communicate it this way is based
on the reason to raise awareness of the change in the
default handling and the stricter restriction for the
field value lenght to match the constraints.

In case systems misused the field and entered invalid
UUID data they need to take care of it, production or
test code.

Resolves: #103141
Releases: main
Change-Id: I8ec94c92d21fa97db9e5fc5170bad42de06e8802
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83026
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Garvin Hicking <garvin@hick.ing>