[!!!][TASK] Throw specific exception in Folder->getSubFolder()
Throwing "global" exception at runtime that are expected to be
caught by calling code is almost always a code smell as outlined
in [1] for a long time already.

We still have various places in the core that violate this,
especially FAL/Resource handling is problematic in this area.

The patch rules out one of such cases:

Folder->getSubFolder() throws a \InvalidArgumentException if the
folder does not exist. There is no hasSubFolder() or similar.
Code thus tends to use an "optimistic" approach: "Get the sub folder,
and if this throws, create it". This forces \InvalidArgumentException
to be caught. This may hide away lower level errors which are
not intended to be caught, which may in turn lead to other issues.

The patch changes getSubFolder() to throw the specific
FolderDoesNotExistException instead, allowing consumers to
handle the "optimistic" approach correctly and without side
effects. Since FolderDoesNotExistException does not inherit from
\InvalidArgumentException, the change has to be considered
breaking. There is no good way around this, so we document this
in a ReST file and outline proper handling within extensions.

[1] https://docs.typo3.org/m/typo3/reference-coreapi/main/en-us/CodingGuidelines/PhpArchitecture/WorkingWithExceptions.html

Resolves: #105920
Related: #96727
Releases: main
Change-Id: I444ab5a51b9fb9fd5403ec403771af769442eb58
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87718
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Mathias Brodala <mbrodala@pagemachine.de>
Reviewed-by: Mathias Brodala <mbrodala@pagemachine.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>