[TASK] Remove collapsing docheader and move to scaffold grid
The collapsing docheader was introduced to make more space
available on smaller screens sizes, but always came at the
drawback of moving parts that could lead to a visual
scissor effect.

This effect needed to be accounted for in several parts
of the backend and was not available in every module,
this is an inconsistency we are now removing. We think
it is better to have a consistant ui than a really small
benefit in space consumption with debatable results.

The docheader is now always sticky and kept in it's
position, to make it fell more like an application
with steady elements that you can rely on reaching
always the same way. We are moving the scrollable
part to the body of the module frame. This allows
us to built better optimized interfaces for certain
situations, as the frame is more stable, like in
the forms module.

The backend is moving away from a flexbox implementation
towards a grid system with defined sections of the
backend. This also allows for outside control of grid areas.

Resolves: #106381
Releases: main, 13.4
Change-Id: I91b6f0d2645abb56e49106b61b5e8163cd6cce75
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88751
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>