[TASK] Have workspace aware inline children
Scenario: A parent table with a type=inline and foreign_table
definition, either as direct DB column, or within a flex form.
The parent table *is* workspace aware (ctrl versioningWS = true),
the child table *is not* workspace aware (ctrl versioningWS = false
or not set).

This scenario leads to a plethora of issues: It starts with
multiple PHP array error access warnings, the refindex does not deal
with it, there are hard logical issues on DB level, and the backend
UI does not deal with it in a sane way by failing to notify workspace
editors they are currently editing live content.

The decision is now: This scenario is not supported. When an inline
parent table is workspace aware, attached "foreign_table" based child
tables have to be workspace aware as well.

The patch adds a TCA migration to scan for this scenario and to make
affected child tables workspace aware automatically. This is always
done, even if the workspace extension is not loaded to prevent
issues when extension authors do not develop with workspaces in mind
but extension users have workspaces loaded.

Notes:
* The change takes care of type=inline with foreign_table only. It does
  not deal with type=group,select,file,category or inline with MM. We
  *may* extend this later, but TYPO3 core has a type=group based relation
  from workspace enabled sys_file_reference to not workspace enabled
  sys_file, which needs to be supported.
* The migration does not detect a workspace aware parent when the
  inline relation is part of a flex form field definition. This is due
  to current restrictions of flex forms and its on-the-fly migration.
  In those relations, a not workspace aware child table is not migrated,
  developers should still make those child tables workspace aware. Further
  refactorings of TCA flex form definitions may however make this scenario
  auto-migratable as well.
* The patch removes the related test scenarios we created with #98103.
  The change is targeted to main, but we may have a v13 change to remove
  those tests as well to prevent useless test fails when dealing with
  issues.

Resolves: #106821
Related: #98103
Releases: main
Change-Id: Ia4fbbb129f434df29525fd45c1cc960e121febb3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88739
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Oli Bartsch <bo@cedev.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>