[BUGFIX] Properly handle errors on page tree processing
When doing page tree processing tasks in styleguide,
there might be some errors, e.g. a missing database
field on creation of the page tree. As this previously
was not properly caught, a Notification with only
the status code "500" was displayed. Additionally, the
button icon was not changed correctly.

This is now fixed. Exceptions on processing are
caught and properly displayed in a notification.

Additionally is the correct icon reset now.

Resolves: #107433
Releases: main, 13.4
Change-Id: Ibf74390f1ed328da8a12f32fc67f41bbe3f703c7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90719
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: Oli Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>