[BUGFIX] Fix dropdowns in element browser
With the modal refactoring to a native dialog element in #107443
the TYPO3 modal implementation dropped it's dependency to the
bootstrap javascript library and therefore the element browser
(which loaded the `@typo3/backend/modal.js` module in modal frames)
did no longer implicitly load bootstrap JavaScript.

That means markup triggers with an implicit bootstrap dependency
did not work. We could now make sure to "always" load the bootstrap
JS module in modal frames, but we take this as an opportunity to
phase out this implicit dependency and instead make sure that components
explicitly load "bootstrap" if they need it.

Releases: main
Resolves: #107829
Related: #107443
Change-Id: I9a2bfe0ab95026db607746b33b2e55ae7295bd7c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91258
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: Oli Bartsch <bo@cedev.de>
Tested-by: Michael Telgkamp <michael.telgkamp@mindscreen.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Michael Telgkamp <michael.telgkamp@mindscreen.de>