[FEATURE] Introduce itemsProcessors over itemsProcFunc
Items processing for check, radio and select-type fields is
centralized and handled more safely with strictly-typed code.

As part of this, a new TCA option "itemsProcessors" is introduced,
as a future replacement for good old "itemsProcFunc", but no
deprecation on the horizon for now.

TCA "itemsProcessors" allow for multiple processors to
be chained, whereas extension authors struggled in the
past that "itemsProcFunc" was only available for the last
extension in loading order to set this option.

Beyond a cool new feature, this patch is a building block
for introducing record list filters.

Resolves: #107889
Releases: main
Change-Id: I1f3d946ea62980ae041e20ca7524840a51e2afd3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91285
Tested-by: Garvin Hicking <garvin@hick.ing>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>